Seperate into diffrent modules and fix race condition
